Add click-through display of feature comparisons
By clicking on a swap icon, show a comparison of the same feature across
projects, instead of all features of one project. This requires upgrades
to the scoring calculations in order to take account that the value may
come from somewhere else than the 'current project', which may not
exist. Also streamline the cards (re)creation.